Restructured project to include "tutorial" and "models"
