+++ Tutorials/vae/convEncoders/10Conv4TConv.1BetaLatentSpace.png
