+++ Tutorials/vae/convEncoders/10Conv4TConv.5BetaLatentSpace.png
