+++ Tutorials/vae/convEncoders/10Conv4TconvLatentSpace.png
