+++ Tutorials/vae/convEncoders/10ConvLatentSpace.png
