+++ Tutorials/vae/convEncoders/4Conv.1BetaLatentSpace.png
