+++ Tutorials/vae/convEncoders/4Conv30BetaLatentSpace.png
