+++ Tutorials/vae/convEncoders/4Conv4BetaLatentSpace.png
