+++ Tutorials/vae/convEncoders/4ConvLatentSpace.png
