--- brokenModels/vaeTest1.py
@@ -1,150 +1,0 @@
-from __future__ import print_function
-import argparse
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-import torch.optim as optim
-from torchvision import datasets, transforms
-from torchsummary import summary
-
-"""
-First attempt at Vanilla Variational Autoencoder
-1 Hidden layer encoder, 1 Hidden layer decoder.
-"""
-
-class Net (nn.Module) :
-
-	def __init__(self):
-		super(Net, self).__init__()
-		
-		#encoder hidden layer
-		self.hiddenEncode1 = nn.Linear(784, 512)
-		
-		#mean and log-variance encoding layers
-		self.mu = nn.Linear(512, 2)
-		self.sigma = nn.Linear(512,2)
-		
-		#decoder hidden layer
-		self.hiddenDecode1 = nn.Linear(2, 512)
-		
-		#output layer
-		self.output = nn.Linear(512, 784)
-		
-	def forward(self, x):
-		#flatten 28-28 image
-		x = x.view(-1, 784)
-		
-		#run input through encoder hidden layer and apply relu
-		x = F.relu(self.hiddenEncode1(x))
-		
-		#get mean and log-variance for input
-		mean = self.mu(x)
-		logvar = self.sigma(x)
-		
-		#calculate standard deviation from log-variance
-		stdDev = torch.exp(.5 * logvar)
-		
-		#get a random variable for pulling from Gaussian distribution
-		epsilon = torch.randn_like(stdDev)
-		
-		#transform random variable with mean and standard deviation to generate code
-		z = epsilon.mul(stdDev).add_(mean)
-		
-		#run code through decoder hidden layer and apply relu
-		z = F.relu(self.hiddenDecode1(z))
-		
-		#run code through output layer and apply sigmoid
-		z = F.sigmoid(self.output(z))
-		
-		#transform output to image dimensions
-		z = z.view(-1, 28, 28)
-		
-		#return output point, mean, and log-variance
-		return z, mean, logvar
-		
-def lossCalc(data, output, mean, logvar):
-	MSE = F.mse_loss(data, output)
-	
-	KLD = -0.5 * torch.sum(1 + logvar - mean.pow(2) - logvar.exp())
-		
-	return MSE + KLD
-	
-
-def train(args,model,device,train_loader,optimizer,epoch):
-	model.train()
-	for batch_idx, (data,target) in enumerate(train_loader):
-		data,target=data.to(device),target.to(device)
-		optimizer.zero_grad()
-		
-		output, mean, logvar = model(data)
-		loss = lossCalc(data, output, mean, logvar)
-		
-		loss.backward()
-		optimizer.step()
-		if batch_idx % args.log_interval ==0:
-			print('Train Epoch: {} [{}/{} ({:.0f}%)]\tLoss: {:.6f}'.format(
-				epoch, batch_idx * len(data), len(train_loader.dataset),
-				100. * batch_idx / len(train_loader), loss.item()))
-def test(args,model,device,test_loader):
-	model.eval()
-	test_loss=0
-	correct=0
-	with torch.no_grad():
-		for data, target in test_loader:
-			data, target = data.to(device), target.to(device)
-			output=model(data)
-			test_loss+=F.nll_loss(output,target,reduction='sum').item()
-			pred=output.argmax(dim=1,keepdim=True)
-			correct+= pred.eq(target.view_as(pred)).sum().item()
-			
-	test_loss/=len(test_loader.dataset)
-	print('\nTest set: Average loss: {:.4f}, Accuracy: {}/{} ({:.0f}%)\n'.format(
-		test_loss, correct, len(test_loader.dataset),
-		100. * correct / len(test_loader.dataset)))
-		
-def main():
-	# Training settings
-	parser = argparse.ArgumentParser(description='PyTorch MNIST Example')
-	parser.add_argument('--batch-size', type=int, default=64, metavar='N',help='input batch size for training (default: 64)')
-	parser.add_argument('--test-batch-size', type=int, default=1000, metavar='N',help='input batch size for testing (default: 1000)')
-	parser.add_argument('--epochs', type=int, default=10, metavar='N',help='number of epochs to train (default: 10)')
-	parser.add_argument('--lr', type=float, default=0.01, metavar='LR',help='learning rate (default: 0.01)')
-	parser.add_argument('--momentum', type=float, default=0.5, metavar='M',help='SGD momentum (default: 0.5)')
-	parser.add_argument('--no-cuda', action='store_true', default=False,help='disables CUDA training')
-	parser.add_argument('--seed', type=int, default=1, metavar='S',help='random seed (default: 1)')
-	parser.add_argument('--log-interval', type=int, default=10, metavar='N',help='how many batches to wait before logging training status')
-	parser.add_argument('--save-model', action='store_true', default=False,help='For Saving the current Model')
-
-	args = parser.parse_args()
-	use_cuda = not args.no_cuda and torch.cuda.is_available()
-	torch.manual_seed(args.seed)
-	device = torch.device("cuda" if use_cuda else "cpu")
-	kwargs = {'num_workers': 1, 'pin_memory': True} if use_cuda else {}
-	train_loader = torch.utils.data.DataLoader(
-		datasets.MNIST('../data', train=True, download=True,
-					transform=transforms.Compose([
-						transforms.ToTensor(),
-						transforms.Normalize((0.1307,), (0.3081,))
-					])),
-		batch_size=args.batch_size, shuffle=True, **kwargs)
-	test_loader = torch.utils.data.DataLoader(
-		datasets.MNIST('../data', train=False, transform=transforms.Compose([
-						transforms.ToTensor(),
-						transforms.Normalize((0.1307,), (0.3081,))
-					])),
-		batch_size=args.test_batch_size, shuffle=True, **kwargs)
-	model = Net().to(device)
-	optimizer = optim.SGD(model.parameters(), lr=args.lr, momentum=args.momentum)
-	
-	summary(model,(1,28,28))
-		
-	for epoch in range(1, args.epochs + 1):
-		train(args, model, device, train_loader, optimizer, epoch)
-		test(args, model, device, test_loader)
-
-	if (args.save_model):
-		torch.save(model.state_dict(),"mnist_cnn.pt")
-	
-
-if __name__ == '__main__':
-    main()
